refactor: move RequestContext around
